perf(frontend): code-split prompt generation UIs, comparison view, and drawers out of the playground chunk
ChatMode/CompletionMode load lazily inside ExecutionItems (agents never render
them), the comparison view loads on first 2-entity selection, and the catalog +
session-inspector drawers load on open. Barrel value re-exports of the split
modules are pruned (playground-ui has no sideEffects config, so a static
re-export would pull them back into every consumer chunk); SessionInspector
buttons are imported per-file for the same reason.

Playground dynamic chunk: 2774 kB -> 2501 kB; ~490 kB now loads on demand.
