+++ web/packages/agenta-playground-ui/src/components/ExecutionItems/index.tsx
@@ -1,4 +1,4 @@
-import {useMemo, useRef} from "react"
+import {lazy, Suspense, useMemo, useRef} from "react"
 
 import {workflowMolecule} from "@agenta/entities/workflow"
 import {executionController} from "@agenta/playground"
@@ -9,9 +9,24 @@ import type {ExecutionHeaderProps} from "../ExecutionHeader"
 import ExecutionHeader from "../ExecutionHeader"
 
 import type {ChatModeProps} from "./assets/ChatMode"
-import ChatMode from "./assets/ChatMode"
 import type {CompletionModeProps} from "./assets/CompletionMode"
-import CompletionMode from "./assets/CompletionMode"
+
+// The prompt-playground generation UIs are code-split: agents never render them, and
+// prompt playgrounds load them behind the same placeholder the pending state shows.
+const ChatMode = lazy(() => import("./assets/ChatMode"))
+const CompletionMode = lazy(() => import("./assets/CompletionMode"))
+
+const ExecutionLoadingPlaceholder = () => (
+    <div className="w-full">
+        <div className="h-[48px] border-0 border-b border-solid border-colorBorderSecondary px-4 py-2 bg-[var(--ag-c-FFFFFF)]">
+            <div className="h-6 w-[220px] rounded bg-[var(--ag-rgba-051729-06)] animate-pulse" />
+        </div>
+        <div className="p-4 flex flex-col gap-3">
+            <div className="h-16 rounded border border-solid border-[var(--ag-rgba-051729-08)] bg-[var(--ag-rgba-051729-02)] animate-pulse" />
+            <div className="h-24 rounded border border-solid border-[var(--ag-rgba-051729-08)] bg-[var(--ag-rgba-051729-02)] animate-pulse" />
+        </div>
+    </div>
+)
 
 export interface PlaygroundGenerationsProps {
     entityId: string
@@ -80,17 +95,7 @@ const PlaygroundGenerations: React.FC<PlaygroundGenerationsProps> = ({
     }
 
     if (isExecutionLoading) {
-        return (
-            <div className="w-full">
-                <div className="h-[48px] border-0 border-b border-solid border-colorBorderSecondary px-4 py-2 bg-[var(--ag-c-FFFFFF)]">
-                    <div className="h-6 w-[220px] rounded bg-[var(--ag-rgba-051729-06)] animate-pulse" />
-                </div>
-                <div className="p-4 flex flex-col gap-3">
-                    <div className="h-16 rounded border border-solid border-[var(--ag-rgba-051729-08)] bg-[var(--ag-rgba-051729-02)] animate-pulse" />
-                    <div className="h-24 rounded border border-solid border-[var(--ag-rgba-051729-08)] bg-[var(--ag-rgba-051729-02)] animate-pulse" />
-                </div>
-            </div>
-        )
+        return <ExecutionLoadingPlaceholder />
     }
 
     return (
@@ -111,19 +116,23 @@ const PlaygroundGenerations: React.FC<PlaygroundGenerationsProps> = ({
                         <AgentGenerationPanel entityId={entityId} />
                     </div>
                 ) : null
-            ) : isChat ? (
-                <ChatMode
-                    entityId={entityId}
-                    renderLastTurnFooter={renderLastTurnFooter}
-                    renderControlsBar={renderControlsBar}
-                />
             ) : (
-                <CompletionMode
-                    entityId={entityId}
-                    withControls
-                    appType={appType}
-                    renderTestsetButton={renderTestsetButton}
-                />
+                <Suspense fallback={<ExecutionLoadingPlaceholder />}>
+                    {isChat ? (
+                        <ChatMode
+                            entityId={entityId}
+                            renderLastTurnFooter={renderLastTurnFooter}
+                            renderControlsBar={renderControlsBar}
+                        />
+                    ) : (
+                        <CompletionMode
+                            entityId={entityId}
+                            withControls
+                            appType={appType}
+                            renderTestsetButton={renderTestsetButton}
+                        />
+                    )}
+                </Suspense>
             )}
         </div>
     )
@@ -131,13 +140,11 @@ const PlaygroundGenerations: React.FC<PlaygroundGenerationsProps> = ({
 
 export default PlaygroundGenerations
 
-// Re-export sub-components (canonical names only)
-export {default as ChatMode} from "./assets/ChatMode"
+// Type-only re-exports for the generation modes; their VALUE exports are deliberately
+// absent — the modes are code-split above (a static re-export would pull them back into
+// every chunk that touches this entry, since the package has no sideEffects config).
 export type {ChatModeProps} from "./assets/ChatMode"
-export {default as ChatTurnView} from "./assets/ChatTurnView"
-export {default as CompletionMode} from "./assets/CompletionMode"
 export type {CompletionModeProps} from "./assets/CompletionMode"
-export {default as ExecutionRow} from "./assets/ExecutionRow"
 export type {ExecutionRowProps} from "./assets/ExecutionRow"
 export {default as GatewayToolAssistantActions} from "./GatewayToolAssistantActions"
 export {default as GatewayToolExecuteButton} from "./GatewayToolExecuteButton"
